2D Array solution
Struggling to find a clean/efficient way to build the table data arrays
for Google Charts API so the given DB data is in the correct spot, and
the arrays are all the same length.   This method requires only two
un-nested while-queries of the database, but then has to again loop
through the created 2D array, all of which would have to happen for each
displayed table.

My next solution will likely involve a meta table or group of tables
that holds the given data for each table so entries with no values will
provide null or zero, instead of being all together skipped.  I think
it's likely there's still a better way to handle this, but for the time
being the meta tables can be updated in real time with triggers in SQL,
or in the script that adds data to the database from a provided data
file.