Customized Layout
To make the project presentation ready, I found and customized a nice
Fluid & Dynamic Bootstrap layout to implement.  Unfortunately Google
Charts doesn't behave well if the window has changed at all since the
initial load, and some chart types are not mobile friendly, but the data
is still available, and the filter tools are still functional.